europe/pyproject.toml
[project]
name = "jobs-europe"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "python-dotenv>=1.0",
]
